+++ ming/metadata.py
@@ -222,7 +222,8 @@ class _ClassManager(metaclass=_CurriedProxyClass):
     _proxy_methods = (
         'get', 'find', 'find_by', 'remove', 'count', 'update_partial',
         'group', 'ensure_index', 'ensure_indexes', 'index_information',  'drop_indexes',
-        'find_and_modify', 'aggregate', 'distinct', 'map_reduce', 'inline_map_reduce',
+        'find_one_and_update', 'find_one_and_replace', 'find_one_and_delete',
+        'aggregate', 'distinct', 'map_reduce', 'inline_map_reduce',
         )
     InstanceManagerClass=_InstanceManager