+++ ming/mim.py
@@ -13,6 +13,7 @@ import warnings
 from datetime import datetime
 from hashlib import md5
 from functools import cmp_to_key
+from enum import Enum
 
 import pickle
 
@@ -330,6 +331,12 @@ class Database(database.Database):
             coll.clear()
 
 
+class ModifyOperation(Enum):
+    UPDATE = 1
+    REPLACE = 2
+    DELETE = 3
+
+
 class Collection(collection.Collection):
     def __init__(self, database, name):
         super().__init__(database, name)
@@ -391,8 +398,8 @@ class Collection(collection.Collection):
             return result
         return None
 
-    def __find_and_modify(self, query=None, update=None, fields=None,
-                          upsert=False, remove=False, **kwargs):
+    def __find_and_modify(self, query=None, update=None, projection=None,
+                          upsert=False, operation=ModifyOperation.UPDATE, **kwargs):
         if query is None: query = {}
         before = self.find_one(query, sort=kwargs.get('sort'))
         upserted = False
@@ -400,49 +407,60 @@ class Collection(collection.Collection):
             upserted = True
             if upsert:
                 result = self.__update(query, update, upsert=True)
-                query = {'_id': result['upserted']}
+                query = {'_id': result['upsert_id']}
             else:
                 return None
 
         before = self.find_one(query, sort=kwargs.get('sort'))
 
-        if remove:
+        if operation == ModifyOperation.DELETE:
             self.__remove({'_id': before['_id']})
-        elif not upserted:
-            self.__update({'_id': before['_id']}, update)
+        else:
+            if operation == ModifyOperation.REPLACE:
+                if any(k.startswith('$') for k in update.keys()):
+                    raise ValueError('replacement can not include $ operators')
+                # FIXME: shouldn't have to do this to mimic mongodb behavior.
+                # But this is all internal, so who cares?
+                update['_id'] = before['_id']
+                self.__update({'_id': before['_id']}, update)
+            elif operation == ModifyOperation.UPDATE:
+                if not all(k.startswith('$') for k in update.keys()):
+                    raise ValueError('update only works with $ operators')
+                if not upserted:
+                    self.__update({'_id': before['_id']}, update)
 
-        return_new = kwargs.get('new', False)
-        if return_new:
-            return self.find_one(dict(_id=before['_id']), fields)
+        return_document = kwargs.pop('return_document', None)
+        return_new = kwargs.pop('new', None)
+        if return_new is not None:
+            warnings.warn('The kwarg new=True is now deprecated. Please use return_document=True instead.', DeprecationWarning, stacklevel=2)
+            if return_document is None:
+                return_document = return_new
+
+        if return_document:
+            return self.find_one(dict(_id=before['_id']), projection)
         elif upserted:
             return None
         else:
-            return Projection(fields).apply(before)
-
-    def find_and_modify(self, query=None, update=None, fields=None,
-                        upsert=False, remove=False, **kwargs):
-        # FIXME: remove
-        # raise NotImplementedError("No longer exists")
-        warnings.warn('find_and_modify is now deprecated, please use find_one_and_delete, '
-                      'find_one_and_replace, find_one_and_update)', DeprecationWarning, stacklevel=2)
-        return self.__find_and_modify(query, update, fields, upsert, remove, **kwargs)
+            return Projection(projection).apply(before)
 
     def find_one_and_delete(self, filter, projection=None, sort=None, **kwargs):
-        return self.__find_and_modify(filter, fields=projection, remove=True, sort=sort, **kwargs)
+        return self.__find_and_modify(filter, projection=projection, operation=ModifyOperation.DELETE, sort=sort, **kwargs)
 
-    def find_one_and_replace(self, filter, replacement, projection=None, sort=None,
-                             return_document=False, **kwargs):
-        # ReturnDocument.BEFORE -> False
-        # ReturnDocument.AFTER -> True
-        return self.__find_and_modify(filter, update=replacement, fields=projection,
-                                      sort=sort, new=return_document, **kwargs)
+    def find_one_and_replace(self, filter, replacement, projection=None, sort=None, upsert=False,
+                             return_document=None, **kwargs):
+        # pymongo.collection.ReturnDocument.BEFORE -> False
+        # pymongo.collection.ReturnDocument.AFTER -> True
+        return self.__find_and_modify(filter, update=replacement, projection=projection,
+                                      sort=sort, upsert=upsert, return_document=return_document, 
+                                      operation=ModifyOperation.REPLACE, **kwargs)
 
-    def find_one_and_update(self, filter, update, projection=None, sort=None,
-                            return_document=False, **kwargs):
-        # ReturnDocument.BEFORE -> False
-        # ReturnDocument.AFTER -> True
-        return self.__find_and_modify(filter, update=update, fields=projection,
-                                      sort=sort, new=return_document, **kwargs)
+    def find_one_and_update(self, filter, update, projection=None, sort=None, upsert=False,
+                            return_document=None, **kwargs):
+        # pymongo.collection.ReturnDocument.BEFORE -> False
+        # pymongo.collection.ReturnDocument.AFTER -> True
+        return self.__find_and_modify(filter, update=update, projection=projection,
+                                      sort=sort, upsert=upsert, return_document=return_document,
+                                      operation=ModifyOperation.UPDATE, **kwargs)
 
     def count(self, filter=None, **kwargs):
         return self.find(filter, **kwargs).count()
@@ -490,24 +508,25 @@ class Collection(collection.Collection):
     def __update(self, spec, updates, upsert=False, multi=False):
         bson_safe(spec)
         bson_safe(updates)
+
+        # https://pymongo.readthedocs.io/en/stable/api/pymongo/results.html#pymongo.results.UpdateResult
         result = dict(
-            connectionId=None,
-            updatedExisting=False,
-            err=None,
-            ok=1.0,
-            n=0,
-            nModified=0
+            acknowleged=True,
+            matched_count=0,
+            modified_count=0,
+            raw_result=None,
+            upsert_id=None,
         )
         for doc, mspec in self._find(spec):
             self._deindex(doc)
             mspec.update(updates)
             self._index(doc)
-            result['n'] += 1
-            result['nModified'] += 1
+            result['matched_count'] += 1
+            result['modified_count'] += 1
             if not multi:
                 break
-        if result['n']:
-            result['updatedExisting'] = True
+        if result['matched_count']:
+            result['acknowleged'] = True
             return result
         if upsert:
             doc = dict(spec)
@@ -519,7 +538,7 @@ class Collection(collection.Collection):
                 raise DuplicateKeyError('duplicate ID on upsert')
             self._index(doc)
             self._data[_id] = bcopy(doc)
-            result['upserted'] = _id
+            result['upsert_id'] = _id
             return result
         else:
             return result